Move census data into separate component (#201)
