+++ frontend/src/Components/AgencyData/AgencyHeader.styled.js
@@ -18,7 +18,7 @@ export const AgencyHeader = styled.div`
   @media (${phoneOnly}) {
     display: flex;
     flex-direction: column;
-    align-items: center;
+    align-items: start;
     justify-content: center;
   }
 `;
@@ -65,16 +65,6 @@ export const EntityDetails = styled.div`
   flex-direction: column;
 `;
 
-export const CensusDemographics = styled.div`
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-
-  @media (${phoneOnly}) {
-    display: none;
-  }
-`;
-
 export const AgencyTitle = styled(H2)`
   font-size: 18px;
 `;
@@ -83,74 +73,11 @@ export const ReportedDate = styled(P)`
   display: inline;
 `;
 
-export const CensusTitle = styled(H2)`
-  font-size: 18px;
-  text-decoration-line: underline;
-  text-decoration-style: dotted;
-  text-underline-offset: 5px;
-  cursor: pointer;
-`;
-
-export const CensusRow = styled.ul`
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  list-style-type: none;
-  overflow-x: auto;
-  overflow-y: hidden;
-  @media (${smallerThanTabletLandscape}) {
-    height: 100px;
-  }
-`;
-
-export const CensusDatum = styled.li`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  &:not(:last-child) {
-    margin-right: 2em;
-  }
-  @media (${smallerThanDesktop}) {
-    flex-direction: row;
-  }
-`;
-
-export const CensusRace = styled(P)`
-  text-transform: uppercase;
-  font-size: 14px;
-  @media (${smallerThanDesktop}) {
-    margin-right: 1em;
-  }
-`;
-
-export const Datum = styled(P)`
-  font-size: 14px;
-`;
-
 export const Other = styled(P)`
   font-size: 14px;
 `;
 
-export const NoCensus = styled.p``;
-
 export const ShowDepartmentsButton = styled.div`
   display: block;
   margin: 10px 0;
-`;
-
-export const Tooltip = styled.div`
-  background: #333;
-  color: white;
-  font-weight: bold;
-  padding: 4px 8px;
-  font-size: 13px;
-  border-radius: 4px;
-  visibility: hidden;
-
-  &[data-show='true'] {
-    visibility: visible;
-  }
 `;
